hiver_aop

Module runtime

AOP Runtime support / AOP 运行时支持

This module provides runtime support for AOP, including:

JoinPoint: Represents a method execution join point

PointcutExpression: Represents a pointcut expression

AspectRegistry: Registers and manages aspects

Proxy: Generates proxies that apply aspects

此模块提供 AOP 的运行时支持，包括：

JoinPoint: 表示方法执行的连接点

PointcutExpression: 表示切点表达式

AspectRegistry: 注册和管理切面

Proxy: 生成应用切面的代理

Structs

AdviceChain

Ordered advice to be applied to a join point. 要应用到连接点的有序通知。

AspectRegistry

Registry for AOP aspects AOP 切面注册表

InterceptChain

Runtime interception chain that applies ordered advice around a target method. 在目标方法周围应用有序通知的运行时拦截链。

InterceptResult

Result of intercepting a method call through the advice chain. 通过通知链拦截方法调用的结果。

JoinPoint

Represents a join point in the program execution 表示程序执行中的连接点

PointcutExpression

Represents a pointcut expression 表示切点表达式

ProceedingJoinPoint

A join point that supports proceeding with the underlying method. 支持继续执行底层方法的连接点。

Enums

AdviceType

Type of advice 通知类型

Functions

global_registry

Get the global aspect registry 获取全局切面注册表